Remove stale snapshots

+++ scripts_debian/custom_wheezy_whonix/01_install_core_post.sh
@@ -30,7 +30,7 @@ fi
 splitPath "$IMG" path_parts
 PREPARED_IMG="${path_parts[dir]}${path_parts[base]}-debootstrap${path_parts[dotext]}"
 
-if ! [ -f "$PREPARED_IMG" ]; then
+if ! [ -f "$PREPARED_IMG" ] && ! [ -f "$INSTALLDIR/tmp/.prepared_whonix" ]; then
     umount_kill "$INSTALLDIR" || :
     warn "Copying $IMG to $PREPARED_IMG"
     cp -f "$IMG" "$PREPARED_IMG"
